Fix toolmanager x/y scale "togglers".
ToolXScale and ToolYScale actually *cannot* be implemented using
ToolToggleBase because a single "toggle" flag cannot store the
log/linear state of *multiple* axes at once (i.e. the behavior of
pressing "l" becomes wrong as soon as there are multiple axes in a given
figure).  Instead these must be plain tools that inspect the actual
scale on the axes before deciding what to do.

In practice, the easiest is to again reuse the classic toolbar
implementation, as is already done for ToolGrid and ToolMinorGrid.
Factor out that "forward-to-classic-toolbar" implementation into a
_ToolForwardingToClassicToolbar base class and use that whereever
possible.