FIX: Checks for (value, color) tuples in from_list
Fixes #29042.

By default, a list of colors is assumed. Only if the conversion to
rgba values fails, check for (value, color) pairs.

The check of the first entry in colors in preserved to raise more
meaningful error messages.

For the first item to be interpreted as a valid (value, color) tuple,
it has to be
 - of type 'Sized' (implements __len__),
 - not of type 'str',
 - length 2.
 - first entry being a real number
 - second entry passing the 'matplotlib.colors.is_color_like' test

Either way, errors are returned if not all entries are (value, color)
pairs or if any invalid colors are encountered.

Implemented tests for different combinations of inputs.