FIX: Error-handling for from_list inputs
Remove complicated check of the first item and always raise both
errors using `raise e2 from e`.

Check for range and monotony of values when unpacking (value, color)
pairs.

Combined tests using `@pytest.mark.parameterize`. Added tests for
values that are not in the range of [0, 1] or that do not increase
monotonically.